Refactor: 커스텀 키보드 context 분리
