+++ src/app/custom-keyboard/_components/canvas/parts/CanvasLoading.tsx
@@ -1,5 +1,7 @@
 import classNames from 'classnames/bind';
+
 import CanvasLoadingIcon from '@/public/svgs/canvasLoading.svg';
+
 import styles from './CanvasLoading.module.scss';
 
 const cn = classNames.bind(styles);
