+++ src/app/custom-keyboard/_components/canvas/parts/StepIcon.tsx
@@ -4,6 +4,7 @@ import classNames from 'classnames/bind';
 
 import CompleteIcon from '@/public/svgs/stepComplete.svg';
 import type { CustomKeyboardStepStatusTypes } from '@/types/CustomKeyboardTypes';
+
 import styles from './StepIcon.module.scss';
 
 const cn = classNames.bind(styles);
