+++ src/app/custom-keyboard/_components/navigator/parts/TooltipColor.tsx
@@ -4,6 +4,7 @@ import classNames from 'classnames/bind';
 
 import { getColorUpperCase } from '@/libs/getColorUpperCase';
 import { convertHexToHsva, convertHsvaToHsla } from '@/libs/convertColor';
+
 import styles from './TooltipColor.module.scss';
 
 const cn = classNames.bind(styles);
