+++ src/app/custom-keyboard/_components/option/Option.tsx
@@ -2,7 +2,7 @@
 
 import { useContext } from 'react';
 
-import { StepContext } from '@/context/customKeyboardContext';
+import { StepContext } from '@/context';
 import BoardOption from './BoardOption';
 import SwitchOption from './SwitchOption';
 import KeyCapOption from './KeyCapOption';
